works

+++ JavaEE/src/main/java/Lesson1/Service/PostService.java
@@ -1,7 +1,12 @@
 package Lesson1.Service;
 
 import Lesson1.DAO.PostDAO;
+import Lesson1.DAO.RelationshipDAO;
+import Lesson1.Exceptions.BadRequestException;
 import Lesson1.Model.Post;
+import Lesson1.Model.Relationship;
+import Lesson1.Model.RelationshipStatus;
+import Lesson1.Model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,26 +17,32 @@ import java.util.List;
 public class PostService {
 
     private final PostDAO dao;
+    private final RelationshipDAO relationshipDAO;
 
     @Autowired
-    public PostService(PostDAO dao) {
+    public PostService(PostDAO dao, RelationshipDAO relationshipDAO) {
         this.dao = dao;
+        this.relationshipDAO = relationshipDAO;
     }
 
-    public Post save(Post post) {
+    public Post save(Post post) throws BadRequestException {
+        validation(post);
         return dao.savePost(post);
     }
 
-    public void delete(long id) {
+    public void delete(long id) throws BadRequestException {
+        Post post = dao.getPost(id);
+        if (post.getUserPosted().getId() != id) throw new BadRequestException("You can't modify this post");
         dao.deletePost(id);
     }
 
-    public Post update(Post post) {
+    public Post update(Post post) throws BadRequestException {
+        validation(post);
         return dao.updatePost(post);
     }
 
     public Post getById(long id) {
-        Post result = dao.findPost(id);
+        Post result = dao.getPost(id);
         if (result == null) throw new EntityExistsException();
         return result;
 
@@ -51,5 +62,14 @@ public class PostService {
 
     public List<Post> getUserPosts(long id) {
         return dao.getUserPosts(id);
+    }
+
+    private void validation(Post post) throws BadRequestException {
+        User userPosted = post.getUserPosted();
+        User userPagePosted = post.getUserPagePosted();
+        Relationship isFriends = relationshipDAO.getRelationship(userPosted.getId(), userPagePosted.getId());
+        if (isFriends == null || isFriends.getStatus() != RelationshipStatus.accepted)
+            throw new BadRequestException("User's not friend");
+        if (post.getMessage().replaceAll("\\s+", "").isEmpty()) throw new BadRequestException("Message is empty");
     }
 }
